Implemented increase endpoint
