--- api/apiserver/pkg/database/models/error.go
@@ -1,37 +1,0 @@
-/*
-Copyright 2021 The PlanetPulse Authors.
-
-Planet Pulse is an API designed to serve climate data pulled from NOAA's
-Global Monitoring Laboratory FTP server. This API is based on the
-OpenAPI v3 specification.
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-A copy of the GNU General Public License can be found here:
-https://www.gnu.org/licenses/
-
-API version: 0.1.0
-Contact: planetpulse.api@gmail.com
-*/
-
-package models
-
-// ErrorResp represents the format of the JSON response returned to the client when an error occurs.
-type ErrorResp struct {
-	Description string
-	Content     ErrorType
-}
-
-// ErrorType represents the context of an error returned to the client.
-type ErrorType struct {
-	Code    int
-	Message string
-}
